services/auth-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyjwt==2.8.0
pydantic==2.9.2

